refactor draft-editor: decomposition
