--- frontend-app/welcome-app/draft-editor.js
@@ -1,190 +1,0 @@
-import React from 'react';
-import { Editor, RichUtils } from 'draft-js';
-import { createEditorStateFromRawDraft} from './helpers/convert-editor-state.js';
-
-class DraftEditor extends React.Component {
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            editorState: createEditorStateFromRawDraft(props.initialRawDraft)
-        };
-
-        this.focus = () => this.refs.editor.focus();
-        this.onChange = (editorState) => this.setState({editorState});
-        this.onSaveDraft = (editorState) => this.props.onSaveDraft(editorState);
-
-        this.handleKeyCommand = (command) => this._handleKeyCommand(command);
-        this.toggleBlockType = (type) => this._toggleBlockType(type);
-        this.toggleInlineStyle = (style) => this._toggleInlineStyle(style);
-        this.saveDraft = (e) => this._saveDraft(e);
-    }
-
-    _saveDraft(e) {
-        this.props.onSaveDraft(this.state.editorState);
-    }
-
-    _handleKeyCommand(command) {
-        const {editorState} = this.state;
-        const newState = RichUtils.handleKeyCommand(editorState, command);
-        if (newState) {
-            this.onChange(newState);
-            return true;
-        }
-        return false;
-    }
-
-    _toggleBlockType(blockType) {
-        this.onChange(
-            RichUtils.toggleBlockType(
-                this.state.editorState,
-                blockType
-            )
-        );
-    }
-
-    _toggleInlineStyle(inlineStyle) {
-        this.onChange(
-            RichUtils.toggleInlineStyle(
-                this.state.editorState,
-                inlineStyle
-            )
-        );
-    }
-
-    render() {
-        const {editorState} = this.state;
-
-        // If the user changes block type before entering any text, we can
-        // either style the placeholder or hide it. Let's just hide it now.
-        let className = 'RichEditor-editor';
-        var contentState = editorState.getCurrentContent();
-        if (!contentState.hasText()) {
-            if (contentState.getBlockMap().first().getType() !== 'unstyled') {
-                className += ' RichEditor-hidePlaceholder';
-            }
-        }
-
-        return (
-            <div>
-                <div className="RichEditor-root">
-                    <BlockStyleControls
-                        editorState={editorState}
-                        onToggle={this.toggleBlockType}
-                    />
-                    <InlineStyleControls
-                        editorState={editorState}
-                        onToggle={this.toggleInlineStyle}
-                    />
-                    <div className={className} onClick={this.focus}>
-                        <Editor
-                            blockStyleFn={this.props.blockStyleFn}
-                            customStyleMap={this.props.customStyleMap}
-                            editorState={editorState}
-                            handleKeyCommand={this.handleKeyCommand}
-                            onChange={this.onChange}
-                            placeholder="Tell a story..."
-                            ref="editor"
-                            spellCheck={true}
-                        />
-                    </div>
-
-                </div>
-                <div>
-                    <input type="button" onClick={this.saveDraft} value="Save Draft" />
-                </div>
-            </div>
-        );
-    }
-}
-
-class StyleButton extends React.Component {
-    constructor() {
-        super();
-        this.onToggle = (e) => {
-            e.preventDefault();
-            this.props.onToggle(this.props.style);
-        };
-    }
-
-    render() {
-        let className = 'RichEditor-styleButton';
-        if (this.props.active) {
-            className += ' RichEditor-activeButton';
-        }
-
-        return (
-            <span className={className} onMouseDown={this.onToggle}>
-              {this.props.label}
-            </span>
-        );
-    }
-}
-
-const BLOCK_TYPES = [
-    {label: 'H1', style: 'header-one'},
-    {label: 'H2', style: 'header-two'},
-    {label: 'H3', style: 'header-three'},
-    {label: 'H4', style: 'header-four'},
-    {label: 'H5', style: 'header-five'},
-    {label: 'H6', style: 'header-six'},
-    {label: 'Blockquote', style: 'blockquote'},
-    {label: 'UL', style: 'unordered-list-item'},
-    {label: 'OL', style: 'ordered-list-item'},
-    {label: 'Code Block', style: 'code-block'},
-];
-
-const BlockStyleControls = (props) => {
-    const {editorState} = props;
-    const selection = editorState.getSelection();
-    const blockType = editorState
-        .getCurrentContent()
-        .getBlockForKey(selection.getStartKey())
-        .getType();
-
-    return (
-        <div className="RichEditor-controls">
-            {BLOCK_TYPES.map((type) =>
-                    <StyleButton
-                        key={type.label}
-                        active={type.style === blockType}
-                        label={type.label}
-                        onToggle={props.onToggle}
-                        style={type.style}
-                    />
-            )}
-        </div>
-    );
-};
-
-var INLINE_STYLES = [
-    {label: 'Bold', style: 'BOLD'},
-    {label: 'Italic', style: 'ITALIC'},
-    {label: 'Underline', style: 'UNDERLINE'},
-    {label: 'Monospace', style: 'CODE'},
-];
-
-const InlineStyleControls = (props) => {
-    var currentStyle = props.editorState.getCurrentInlineStyle();
-    return (
-        <div className="RichEditor-controls">
-            {INLINE_STYLES.map(type =>
-                    <StyleButton
-                        key={type.label}
-                        active={currentStyle.has(type.style)}
-                        label={type.label}
-                        onToggle={props.onToggle}
-                        style={type.style}
-                    />
-            )}
-        </div>
-    );
-};
-
-DraftEditor.propTypes = {
-    blockStyleFn: React.PropTypes.func.isRequired,
-    customStyleMap: React.PropTypes.object.isRequired,
-    initialRawDraft: React.PropTypes.object.isRequired,
-    onSaveDraft: React.PropTypes.func.isRequired
-};
-export default DraftEditor;
