+++ frontend-app/welcome-app/editor/header/inline-style-controls.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import StyleButton from './style-button.js';
+import { INLINE_STYLES } from './inline-style-types.js';
+
+export const InlineStyleControls = (props) => {
+    var currentStyle = props.editorState.getCurrentInlineStyle();
+    return (
+        <div className="RichEditor-controls">
+            {INLINE_STYLES.map(type =>
+                    <StyleButton
+                        key={type.label}
+                        active={currentStyle.has(type.style)}
+                        label={type.label}
+                        onToggle={props.onToggle}
+                        style={type.style}
+                    />
+            )}
+        </div>
+    );
+};
+
