+++ frontend-app/welcome-app/editor/header/inline-style-types.js
@@ -1,0 +1,6 @@
+export const INLINE_STYLES = [
+    {label: 'Bold', style: 'BOLD'},
+    {label: 'Italic', style: 'ITALIC'},
+    {label: 'Underline', style: 'UNDERLINE'},
+    {label: 'Monospace', style: 'CODE'},
+];
